Update function delete promotion

+++ public/frontend/js/list_promotion.js
@@ -23,20 +23,34 @@ $(document).ready(function(){
     table.on('click', 'button.delete', function(e){
         var id_delete = $(this).val();
         var url_del = $('#url_delete').val() + '/' + id_delete;
-        $.ajax({
-            url : url_del,
-            type : 'DELETE',
-            dataType: 'json',
-            success : function (result) {
-            	table.clear().draw();
-                $('#message').html(result.message);
-	            $('#message').css('display', 'block');
-            },
-            error : function (data) {
-	            var err = eval("(" + data.responseText + ")");
-	            $('#message').html(err.error);
-	            $('#message').css('display', 'block');
-            }
+        swal({
+            title: messages.confirm_delete_title,
+            text: messages.confirm_delete_promotion,
+            type: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: messages.delete,
+            closeOnConfirm: false
+        }, function () {
+            $.ajax({
+	            url : url_del,
+	            type : 'DELETE',
+	            dataType: 'json',
+	            success : function (result) {
+	            	table.clear().draw();
+	            	swal({
+                        title: result.message,
+                        confirmButtonColor: "#DD6B55",
+                        confirmButtonText: messages.ok,
+                    });
+	            },
+	            error : function (data) {
+		            var err = eval("(" + data.responseText + ")");
+		            $('#message').html(err.error);
+		            $('#message').css('display', 'block');
+	            }
+	        });  
         });
+        
     });
 });
